+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [0.1.5] - 2024-03
 ### Added
 - Right click menu
+- Pause option
 
 ### Fixed
 - Update plan bug